Rename parameter object
